requirements.txt
streamlit
groq
pyngrok
pandas
python-dotenv
validators
